\name{divideTwDEMCBatch.divideTwDEMCBatch}
\alias{divideTwDEMCBatch.divideTwDEMCBatch}
\title{divideTwDEMCBatch divideTwDEMCBatch}
\description{append runs to result of \code{\link{divideTwDEMCBatch.default}}}
\usage{
\method{divideTwDEMCBatch}{divideTwDEMCBatch}(x, ...)
}
\arguments{
  \item{x}{result of former call to divideTwDEMCBatch}
  \item{\dots}{further arguments to \code{\link{divideTwDEMCBatch.default}}}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





