\name{getNParms.twDEMC}
\alias{getNParms.twDEMC}
\title{getNParms twDEMC}
\description{Extracts the number of parameters, i.e. the dimension of the parameter vector}
\usage{
\method{getNParms}{twDEMC}(res, ...)
}
\arguments{
  \item{res}{object of class twDEMC}
  \item{\dots}{
}
}

\value{integer, number of parameters in twDEMC}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{getNGen.twDEMC}}
\code{\link{subChains.twDEMC}}
,\code{\link{twDEMCInt}}}

